feat: update the setting menu to comply with WordPress
