+++ inc/classes/class-assets.php
@@ -100,6 +100,20 @@ class Assets {
 			)
 		);
 
+		$godam_settings = get_option( 'rt-easydam-settings' );
+
+		$selected_brand_image = $godam_settings['general']['selected_brand_image'] ?? '';
+		$brand_color          = $godam_settings['general']['brand_color'] ?? '';
+
+		wp_localize_script(
+			'godam-script',
+			'godamSettings',
+			array(
+				'brandImage' => $selected_brand_image,
+				'brandColor' => $brand_color,
+			)
+		);
+
 		wp_enqueue_script( 'godam-script' );
 		wp_enqueue_style( 'godam-style' );
 
@@ -139,6 +153,20 @@ class Assets {
 				'url' => get_rest_url( get_current_blog_id() ),
 				'home_url' => get_home_url( get_current_blog_id() ),
 				'nonce' => wp_create_nonce( 'wp_rest' ),
+			)
+		);
+
+		$godam_settings = get_option( 'rt-easydam-settings' );
+
+		$selected_brand_image = $godam_settings['general']['selected_brand_image'] ?? '';
+		$brand_color          = $godam_settings['general']['brand_color'] ?? '';
+
+		wp_localize_script(
+			'godam-script',
+			'godamSettings',
+			array(
+				'brandImage' => $selected_brand_image,
+				'brandColor' => $brand_color,
 			)
 		);
